[install] Remove previous Xamarin.Android.Sdk.{props,targets} (#1776)
Update the `make install` target so that `Xamarin.Android.Sdk.props`
and `Xamarin.Android.Sdk.target` are always removed.  If these are
not removed, it's possible that older versions from a previous
`make install` invocation will be present, resulting in an
inconsistent installation and related confusion, such as a wrong
`$(TargetFrameworkRootPath)` value.