[Xamarin.Android.Build.Tasks] AndroidCodegenTarget=XAJavaInterop1 (#1774)
When `$(AndroidCodegenTarget)`=XAJavaInterop1, `generator` output uses
`Java.Interop.JniPeerMembers` for JNI member invocations instead of
the legacy `Android.Runtime.JNIEnv` methods.

This has a number of benefits, such as:

  * Reduced memory use (less garbage generated)
  * Smaller assemblies.
  * Faster `base.Method()` invocations.
  * Improved `jmethodID` caching of Java Callable Wrapper constructors
    for type subclasses.

The problem?  Assemblies built under `XAJavaInterop1` cannot run on
Xamarin.Android 6.0 and earlier.

Xamarin.Android 6.0 was over 2.5 *years* ago.  We do not consider this
to be a signifiant concern, while the *memory and performance
benefits* are significant.

Update the default `$(AndroidCodegenTarget)` value to `XAJavaInterop1`.

If this is undesirable, projects can set the `$(AndroidCodegenTarget)`
MSBuild property to `XamarinAndroid` for the previous behavior.