[Xamarin.Android.Build.Tasks] add a SLN to work on build tasks easier
Currently working with Xamarin.Android.Build.Tasks in Visual Studio
for Windows requires several non-working projects to be unloaded.

Most of the development done on Windows involve the
Xamarin.Android.Build.Tasks MSBuild tasks and unit tests, so it makes
sense to just have a simple SLN file to work on this specific part of
the product.

This simplifies the workflow on Windows:
- Run `msbuild Xamarin.Android.sln /t:Prepare`
- Run `msbuild Xamarin.Android.sln`
- Open `Xamarin.Android.Build.Tasks.sln`, which only loads 4 relevant
  projects, and work completely in the IDE from there. The solution
  also loads very quickly, due to having less projects.

From here, the IDE's NUnit unit test runner works, and it is
straightforward to make changes to MSBuild tasks.

Other changes:
- Added a `.gitignore` rule for VS Code