[Security] Fix Zip-Slip vulnerability
Update LibZipSharp to pull in fixes for the zip-slip
vulnerability (https://snyk.io/research/zip-slip-vulnerability).

The fix is to make sure that no zip entry path can be used to create files
outside the designated output directory. This is done by removing both the
relative path segments as well as making sure that the entry path is not rooted.