[Xamarin.Android.Build.Tasks] [Aapt2] Allow dashes in Resource file names
Fixes #1780

We forgot to include `-` as a valid character for the Resource
filenames when implementing the `CheckForInvalidResourceFileNames`
Task. As a result many working samples, when using `aapt2`, failed
with.

	APT0000: Invalid file name: It must contain only [a-zA-Z0-9_.].

The fix is to allow `-` as a filename character.
This is only done for `values-*` resources, since they
are the only ones which support `-`. For all other
resource types `-` is invalid.