Bump to libZipSharp/master/d6951493 (#1788)
Context: https://snyk.io/research/zip-slip-vulnerability

The fix is to make sure that no zip entry path can be used to create
files outside the designated output directory.  This is done by
removing both the relative path segments as well as making sure that
the entry path is not rooted.