Android P API enumification and stabilization.
git diff on remaining-int-*.txt shows what remains (and many of them are
covered by comments on enumification-helpers).

(There should be up-to-date documentation on how to work on enumification
so that anyone can take over the work even after I left the project.)

The API is now marked as stable.