[Xamarin.Android.Build.Tasks] Fix inconsistent library cache files. (#1818)
Fixes: https://github.com/xamarin/xamarin-android/issues/1814

Both of the `GetImportedLibraries` and
`ResolveLibraryProjectImports` targets have a problem where the cache
files they generate are not always consistent between builds: in some
cases, additional entries are added or removed:

	diff -u old.cache Bug15162/obj/Debug/libraryprojectimports.cache
	--- old.cache	2018-06-14 11:12:10.000000000 +0100
	+++ Bug15162/obj/Debug/libraryprojectimports.cache	2018-06-14 11:12:28.000000000 +0100
	@@ -1,8 +1,6 @@
	 ﻿<?xml version="1.0" encoding="utf-8"?>
	 <Paths>
	   <Jars>
	-    <Jar>obj/Debug/lp/0/jl/formsviewgroup.jar</Jar>
	-    <Jar>obj/Debug/lp/7/jl/support-annotations.jar</Jar>
	     <Jar>/Users/dean/Downloads/Bug15162-1/Bug15162/obj/Debug/lp/0/jl/formsviewgroup.jar</Jar>
	     <Jar>/Users/dean/Downloads/Bug15162-1/Bug15162/obj/Debug/lp/10/jl/bin/classes.jar</Jar>
	     <Jar>/Users/dean/Downloads/Bug15162-1/Bug15162/obj/Debug/lp/11/jl/bin/classes.jar</Jar>

	diff -u old.cache Bug15162/obj/Debug/libraryimports.cache
	--- old.cache	2018-06-14 11:29:41.000000000 +0100
	+++ Bug15162/obj/Debug/libraryimports.cache	2018-06-14 11:30:42.000000000 +0100
	@@ -3,27 +3,40 @@
	   <ManifestDocuments>
	     <ManifestDocument>obj/Debug/lp/10/jl/AndroidManifest.xml</ManifestDocument>
	     <ManifestDocument>obj/Debug/lp/11/jl/AndroidManifest.xml</ManifestDocument>
	+    <ManifestDocument>obj/Debug/lp/11/jl/manifest/AndroidManifest.xml</ManifestDocument>
	     <ManifestDocument>obj/Debug/lp/12/jl/AndroidManifest.xml</ManifestDocument>

As a result targets like `_UpdateAndroidResgen` will run when they
really do not need to, increasing build times.

For the `ResolveLibraryProjectImports` target, the extra `jars.Add()`
was not need because at the end of the task we add *all* of the jar
files we find within the output directory.  The additional
`jars.Add()` calls were responsible for adding the non-fullpath
variants of the files.

The `GetImportedLibraries` target was adding the files from a
`manifest` subdirectory in the `lp` intermediate directories, which
likewise isn't needed.