[Xamarin.Android.Buiild.Tasks] Allow uncompressed native libraries in apk
Fixes #1822

Currently we only apply the `$(AndroidStoreUncompressedFileExtensions)` to
the resources added to the apk via `aapt2`. All the other resources added
by `BuildApk` are added as `Deflate`. As a result this means an
Xamarin.Android apk cannot run on `Android.Things` since it requires the
native libraries to be uncompressed.

We should we using the `$(AndroidStoreUncompressedFileExtensions)` property
as part of the `BuildApk` process to ensure that files which need to
be left uncompressed are.