[Xamarin.Android.Build.Tasks] fix for android:name="$" in manifest
Fixes: https://github.com/xamarin/xamarin-android/issues/1698

In the project reproducing this issue, the Xamarin.Firebase.lid NuGet
produces an `AndroidManifest.xml` with the following line:

    <permission android:name="${applicationId}.permission.C2D_MESSAGE" android:protectionLevel="signature" />

Unfortunately our manifest merging code is throwing an exception and
producing a warning such as:

    Xamarin.Android.Common.targets(1910,3): warning : Name cannot begin with the '$' character, hexadecimal value 0x24.

We also aren't getting the contents of this `AndroidManifest.xml`
merged, so this is definitely an issue!

The exception is coming from the following code:

    var existing = (name != null) ?
        doc.XPathSelectElement (string.Format ("/manifest/{0}[@android:name='{1}']", top.Name.LocalName, XmlConvert.VerifyNCName (name.Value)), nsResolver) :
        doc.XPathSelectElement (string.Format ("/manifest/{0}", top.Name.LocalName));

Apparently `XmlConvert.VerifyNCName` throws an exception if the input
value contains a `$`.

It appears it has been this way for quite some time, prior to becoming
open source:

https://github.com/xamarin/xamarin-android/blame/bb8183c8c16667b18ba9197e78294b17a0af532c/src/Xamarin.Android.Build.Tasks/Utilities/ManifestDocument.cs#L444

I suspect the original goal was to validate the XPath expression
wouldn't throw an exception, but it seems to work fine to use the
value as-is. Let's just not call `XmlConvert.VerifyNCName` at all.

After the fix, I get additional lines in my
`obj/Debug/android/AndroidManifest.xml`:

    <permission android:name="com.companyname.DollarSignProb.permission.C2D_MESSAGE" android:protectionLevel="signature" />
    <uses-permission android:name="com.companyname.DollarSignProb.permission.C2D_MESSAGE" />

So it is now properly merging the `AndroidManifest.xml`.

Other changes:
- Updated `ManifestTest.MergeLibraryManifest` test so it also verifies
  this fix
- Added `XA4302` as part of the effort towards:
  https://github.com/xamarin/xamarin-android/issues/1560