[Xamarin.Android.Buiild.Tasks] Allow uncompressed native libraries in apk (#1827)
Fixes: https://github.com/xamarin/xamarin-android/issues/1822

Currently we only apply the
`$(AndroidStoreUncompressedFileExtensions)` MSBuild property to the
resources added to the apk via `aapt2`.  All the other resources
added by the `<BuildApk/>` task are added as `Deflate`.  As a result
this means an Xamarin.Android apk cannot run on `Android.Things`
since it requires the native libraries to be uncompressed.

We should we using the `$(AndroidStoreUncompressedFileExtensions)`
property as part of the `<BuildApk/>` task to ensure that files which
need to be left uncompressed *are* left uncompressed.