[Xamarin.Android.Build.Tasks] Add `FrameworkDirectories` to `GenerateJavaStubs`.
Context https://devdiv.visualstudio.com/DevDiv/_workitems/edit/631540

The Cecil `DirectoryAssemblyResolver` is failing to resolve
`System.Runtime.dll` when loading assebmlies from the Intermediate
directory.

	MSB4018: System.IO.FileNotFoundException: Could not load assembly 'System.Runtime, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a'. Perhaps it doesn't exist in the Mono for Android profile?

This seems to be becaue we have netstandard2.0 references in the
list of assemblies which reference `System.Runtime.dll`. However
in `GenerateJavaStubs` we are not including the `v1.0` or the
`Facades` directory in the search path. As a result the dependency
cannot be loaded. By default `DirectoryAssemblyResolver` only
uses the directories from the assemblies being loaded to search
for dependencies. Because `System.Runtime.dll` is a `Framework`
assembly, it will NEVER be copied into the intermediate directory.

So lets tell `DirectoryAssemblyResolver` about where the other
directories are.