[Xamarin.Android.Build.Tasks] Use `InMemory` for `GenerateResourceDesigner`.
Context https://devdiv.visualstudio.com/DevDiv/_workitems/edit/616623

We seem to be getting a lock in VS when restoring Nuget Packages.
The call stack seems to indicate that we are hanging in
`GenerateResourceDesigner`. Specifically in the `ReadByes` call that
Cecil finally ends up in.

One theory is we are locking the file.. somehow. So just to be sure
we should make sure Cecil is using `InMemory` for the designer.