[tests] kill adb when Xamarin.Android.Build.Tests complete (#1844)
Context: https://devdiv.visualstudio.com/DevDiv/_build?buildId=1768391
Context: https://devdiv.visualstudio.com/DevDiv/_build?buildId=1778620

There are a couple issues happening on VSTS:

 1. If the build needs to delete/recreate `~\android-toolchain\sdk`,
    sometimes `adb.exe` is running and so the directory cannot be
    deleted.
 2. Sometimes the `<Exec/>` MSBuild task will hang indefinitely,
    even though the NUnit tests have completed.

Our test run `adb.exe` here, to determine if a device/emulator is
connected:

https://github.com/xamarin/xamarin-android/blob/103b5a755c048c6eaedb43c139cac32b920372cd/src/Xamarin.Android.Build.Tasks/Tests/Xamarin.Android.Build.Tests/Utilities/BaseTest.cs#L19-L21

It turns out the two issues are directly related. If you launch
`adb.exe`, and it is the first instance of `adb`, it will continue to
run as a daemon-like process.  The `<Exec/>` task will block until
all child processes of `nunit3-console.exe` exit, and so we get a
hang running our tests...

To fix this, I setup a `[SetUpFixture]`, which is a feature of NUnit
that lets you run code before and after all the tests complete. I
moved the `static ctor` of `BaseTest` into a `[OneTimeSetUp]` method,
and then created a `[OneTimeTearDown]` method to kill `adb`.

Other changes:

  - I let VS sort the using statements, since they were oddly ordered.