[build] Split out bundle download from xa-prep-tasks
A minor SNAFU was observed earlier this week: *all* macOS PR builds
were [failing][0] because `NuGet.exe` couldn't be downloaded:

	build-tools/xa-prep-tasks/xa-prep-tasks.targets(91,5): error :
	Unable to download URL `https://dist.nuget.org/win-x86-commandline/latest/nuget.exe` to `/Users/builder/jenkins/workspace/xamarin-android-pr-builder/xamarin-android/build-tools/xa-prep-tasks/../../.nuget/NuGet.exe`:
	An error occurred while sending the request

What we believe happened was as follows:

 1. A [PR to use mono:2018-06 was built][1], which
    [updated the *system* mono to 5.16.0 (also from mono:2018-06)][2]:

        Downloading `https://xamjenkinsartifact.azureedge.net/build-package-osx-mono/2018-06/7/7627a5f9eeba0fd846731ad0c498556f55be1a34/MonoFramework-MDK-5.16.0.5.macos10.xamarin.universal.pkg` to `/Users/builder/android-archives/.MonoFramework-MDK-5.16.0.5.macos10.xamarin.universal.pkg.download`.
        sudo installer -pkg "/Users/builder/android-archives/MonoFramework-MDK-5.16.0.5.macos10.xamarin.universal.pkg" -target /

 2. This particular mono had a "broken" `HttpClient`, in that it
    couldn't download `https://dist.nuget.org/win-x86-commandline/latest/nuget.exe`

 3. All following PR builds that ran on xam-mac-mini-6 would fail, as
    mono 5.16.0 would fail to download `nuget.exe`, causing the
    overall build to fail.

"Funnily enough," this is more or less why we have the
`$(MonoRequiredMaximumVersion)` MSBuild property; see 2bf62cb9.  The
problem *here* is that the download failure and subsequent build
failure happened *before* the mono version was checked, meaning our
build system code to check the mono version was ~useless.

Ouch.

Thus, the *real* problem is that our dependency checks were being
performed *after* we attempted to download anything.  If instead we
checked the mono version *before* we tried to download files, we would
have installed a "known good" mono, and everything would have worked.

This in turn means that `build-tools/xa-prep-tasks` should *not*
perform any downloads, as `xa-prep-tasks` is a dependency for building
`build-tools/dependencies`.

Fix our build ordering problem by moving the download logic out of
`build-tools/xa-prep-tasks` and into a new
`build-tools/download-bundle` project.  This will allow xa-prep-tasks
to build (which is necessary), after which `build-tools/dependencies`
can sanity-check the build environment, and only *after* that
validation is performed will we attempt to download anything and
continue the normal build.

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android-pr-builder/3424/
[1]: https://github.com/xamarin/xamarin-android/pull/1836
[2]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android-pr-builder/3373/