[android-toolchain] Install Android platform-tools r28 (#1858)
If you are running Android O or above as the Android target to test
`make run-apk-tests`, it will fail because of an `adb` server version
mismatch (39 vs. 40).  This is a common `adb` issue e.g.:

https://sqa.stackexchange.com/questions/20524/how-to-resolve-adb-server-version-32-doesnt-match-this-client-36-killin

Bump to Android platform-tools r28 so that we have an updated `adb`,
which will avoid the version mismatch error.