[Xamarin.Android.Build.Tasks] fixes for invalid __AndroidLibraryProjects__.zip (#1852)
Context: https://developercommunity.visualstudio.com/content/problem/218603/msbuild-failure-in-resolvelibraryprojectimports-ta.html
Context: https://github.com/AArnott/PCLCrypto/issues/138

Fixes: http://work.devdiv.io/631267

Developers were reporting an error when consuming the PCLCrypto
`2.1.17-alpha-g5b1e8dff8c` NuGet package:

        error MSB4018: The "ResolveLibraryProjectImports" task failed unexpectedly.
        error MSB4018: System.NotSupportedException: The given path's format is not supported.
        error MSB4018: at System.Security.Permissions.FileIOPermission.EmulateFileIOPermissionChecks(String fullPath)
        error MSB4018: at System.IO.Directory.InternalCreateDirectoryHelper(String path, Boolean checkHost)
        error MSB4018: at System.IO.Directory.CreateDirectory(String path)
        error MSB4018: at Xamarin.Android.Tools.Files.ExtractAll(ZipArchive zip, String destination, Action`2 progressCallback, Func`2 modifyCallback, Func`2 deleteCallback, Boolean forceUpdate)
        error MSB4018: at Xamarin.Android.Tasks.ResolveLibraryProjectImports.Extract(DirectoryAssemblyResolver res, ICollection`1 jars, ICollection`1 resolvedResourceDirectories, ICollection`1 resolvedAssetDirectories, ICollection`1 resolvedEnvironments)
        error MSB4018: at Xamarin.Android.Tasks.ResolveLibraryProjectImports.Execute()
        error MSB4018: at Microsoft.Build.BackEnd.TaskExecutionHost.Microsoft.Build.BackEnd.ITaskExecutionHost.Execute()

Upon trying this myself, I extracted the contents of
`__AndroidLibraryProjects__.zip` and encountered some *very* odd files
paths inside the zip:

        Archive:  __AndroidLibraryProjects__.zip
        Zip file size: 1939 bytes, number of entries: 7
        defN 18-Mar-10 10:12 library_project_imports/__res_name_case_map.txt
        stor 18-Mar-10 10:12 library_project_importsC:/projects/pclcrypto/src/.obj/pclcrypto/Release/monoandroid23/jl/assets/
        stor 18-Mar-10 10:12 library_project_importsC:/projects/pclcrypto/src/.obj/pclcrypto/Release/monoandroid23/jl/bin/
        stor 18-Mar-10 10:12 library_project_importsC:/projects/pclcrypto/src/.obj/pclcrypto/Release/monoandroid23/jl/java/
        stor 18-Mar-10 10:12 library_project_importsC:/projects/pclcrypto/src/.obj/pclcrypto/Release/monoandroid23/jl/res/
        stor 18-Mar-10 10:12 library_project_importsC:/projects/pclcrypto/src/.obj/pclcrypto/Release/monoandroid23/jl/res/values/
        defN 18-Mar-10 10:12 library_project_importsC:/projects/pclcrypto/src/.obj/pclcrypto/Release/monoandroid23/jl/res/values/strings.xml
        7 files, 167 bytes uncompressed, 117 bytes compressed:  29.9%

Since these file paths contain a `:`, we are getting
`System.NotSupportedException` when calling
`Directory.CreateDirectory()`. They are also completely incorrect paths!

Next, I looked at PCLCrypto's project on Github and was able to
publicly find the binary MSBuild log from this release:
https://ci.appveyor.com/project/AArnott/pclcrypto/build/2.1.17-alpha

        CreateManagedLibraryResourceArchive Task
          OutputDirectory: C:\projects\pclcrypto\src\../obj/pclcrypto/Release\monoandroid23\jl
          ResourceDirectory: C:\projects\pclcrypto\src\../obj/pclcrypto/Release\monoandroid23\res\
          AndroidAssets:
          AndroidJavaSources:
          AndroidJavaLibraries:
          AndroidResourcesInThisExactProject:
            C:\projects\pclcrypto\obj\pclcrypto\Release\monoandroid23\res\values\strings.xml
          RemovedAndroidResourceFiles:
        writing __res_name_case_map.txt...
        Saving contents to C:\projects\pclcrypto\obj\pclcrypto\Release\monoandroid23\__AndroidLibraryProjects__.zip

Somehow passing in these odd paths to the
`CreateManagedLibraryResourceArchive` MSBuild task is creating a
malformed `__AndroidLibraryProjects__.zip`!

I cloned the source, and I could reproduce the behavior with VS
2017 15.7.3. Looking at the project's `.csproj` and other MSBuild
files, I found that `$(BaseIntermediateOutputPath)` was declared in
an odd way:

        <BaseIntermediateOutputPath>$(MSBuildThisFileDirectory)../obj/$(MSBuildProjectName)/</BaseIntermediateOutputPath>

So this explains how we are getting
`C:\projects\pclcrypto\src\../obj/pclcrypto/Release\monoandroid23\jl`
as input during the build.  I submitted a PR to PCLCrypto with a
workaround: use Windows path separators.

As for the fix in Xamarin.Android, we really have two issues to fix:

  - Paths such as
    `C:\projects\pclcrypto\src\../obj/pclcrypto/Release\monoandroid23\jl`
    should not produce an invalid `__AndroidLibraryProjects__.zip`
  - We should provide a better error message if an invalid zip is
    encountered during the `<ResolveLibraryProjectImports/>` task.

~~ CreateManagedLibraryResourceArchive ~~

The first fix was simple: there were a few places where
`OutputDirectory` was used as-is, but most of this task is using a
`DirectoryInfo.FullName` value. We can switch to using `FullName`,
since it is the full path and will have its path separators
normalized.

I added the `AndroidLibraryProjectsZipWithOddPaths` test to validate
this fix.

~~ ZipArchiveEx ~~

In addition to the fixes in the
`<CreateManagedLibraryResourceArchive/>` task, it appears that
`ZipArchiveEx` has some weird logic working with paths:

        var internalDir = dir.Replace ("./", string.Empty).Replace (folder, string.Empty);
        string fullDirPath = folderInArchive + internalDir;

Not only is `./` likely ignoring Windows directory separators, but
there is some weird behavior caused by a test case.

If I pass in `folder` as:

        "C:\Users\myuser\Desktop\Git\xamarin-android\bin\TestDebug\temp\../temp/"

Then `internalDir` evaluates to:

        "C:\Users\myuser\Desktop\Git\xamarin-android\bin\TestDebug\temp\.temp/B"

Then `fullDirPath` is:

        "tempC:\Users\myuser\Desktop\Git\xamarin-android\bin\TestDebug\temp\.temp/B"

So the paths here aren't right at all!

I added a new set of `ZipArchiveExTests` to reproduce the issue.

Next, I added some logic to normalize directory separators and call
`Path.GetFullPath` on the incoming folder.

I also added a check to see if the path ends with
`Path.DirectorySeparatorChar`, as that would break this line, which
assumes `internalDir` is prefixed with the directory separator:

        string fullDirPath = folderInArchive + internalDir;

~~ ResolveLibraryProjectImports ~~

As for `<ResolveLibraryProjectImports/>`, it was calling into
`Files.ExtractAll()` where `Directory.CreateDirectory()` calls could
possibly encounter invalid paths that would throw
`System.NotSupportedException`.  We can handle this exception and
makesure the zip entry is reported in the exception.  All we can do
is `throw`, but at least including the errant zip entry text is an
improvement.

Since `<ResolveLibraryProjectImports/>` was already handling
`PathTooLongException`, I changed it to also handle
`NotSupportedException` in the same way.  I created a new `XA4303`
error code along with documentation, to properly report the failure
extracting `__AndroidLibraryProjects__.zip`.

I added the `ResolveLibraryProjectImportsWithInvalidZip` test to
validate these changes.

Other changes:

  - In `Files.ExtractAll()`, it looked like `progressCallback` wasn't
    being invoked at the right time. I moved it at the top of the
    `foreach` loop, so callers will get progress updates for files
    that are skipped.