[tests] fix for BuildAppWithManagedResourceParserAndLibraries timeout
Context: https://github.com/nunit/docs/wiki/Framework-Parallel-Test-Execution#specifying-parallelism-at-multiple-test-levels
Context: http://xamarin-jenkins/job/monodroid-pr/397/

On Jenkin (monodroid builds) a test can fail with:

    1) Failed : Xamarin.Android.Build.Tests.AndroidUpdateResourcesTest.BuildAppWithManagedResourceParserAndLibraries
        DesignTime build should be less than 10000 milliseconds.
        Expected: less than or equal to 10000.0d
        But was:  12820.0d

But this test is marked with `[NonParallelizable]`, what is going on?

Looking at NUnit's documentation, we have this test configured as:

    Parallel class with non-parallel methods: The methods run
    sequentially, usually on the same thread that ran the class one-time
    setup, but may actually be running in parallel with other, unrelated
    methods from other classes.

Since the test class has `[Parallelizable (ParallelScope.Children)]`,
if we mark `BuildAppWithManagedResourceParserAndLibraries` as
`[NonParallelizable]` it can still run in parallel with methods of
other test classes!

So the only way to fix this is:
- Make the test class `[NonParallelizable]`
- Mark each method `[Parallelizable]`, which defaults to a scope of
  `ParallelScope.Self`

Then we get the configuration we want in this case:

    Non-parallel class with parallel methods: The methods only run in
    parallel with one another, not with the test methods of any other
    classes.