[Xamarin.Android.Build.Tasks] Add Warnings for minimum Google Play requirements.
Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/635168

Google as of August and November 2018 will required a minimum
of TargetFrameworkVersion v8.0 to publish/update apps.

This commit adds a new check to `ResolveSdksTask` which will
emit warnings when we hit those dates. This will only happen
if the `TargetFrameworkVersion` < `v8.0`

[1] [Google Play's target API level requirement](https://developer.android.com/distribute/best-practices/develop/target-sdk)