[Xamarin.Android.Build.Tasks] _CleanIntermediateIfNuGetsChange deletes build.props
Context: https://github.com/xamarin/xamarin-android/pull/1878

The initial `_CleanIntermediateIfNuGetsChange` target had a terrible
bug.

If you were building a project for the first time, due to the ordering
of targets:

    <BuildDependsOn>
        _ValidateLinkMode;
        _SetupDesignTimeBuildForBuild;
        _CreatePropertiesCache;
        _AddAndroidDefines;
        _AddNativeLibraryArchiveToCompile;
        _CheckTargetFramework;
        _RemoveLegacyDesigner;
        _ValidateAndroidPackageProperties;
        _CleanIntermediateIfNuGetsChange;
        $(BuildDependsOn);
    </BuildDependsOn>

`_CreatePropertiesCache` creates `build.props`, but then on a first
build `_CleanIntermediateIfNuGetsChange` runs and deletes it!

The fix here is to move `_CleanIntermediateIfNuGetsChange` ahead of
`_CreatePropertiesCache`--it does not appear that any other
intermediate files are created before this target.

I also updated the tests that validate
`_CleanIntermediateIfNuGetsChange` to make sure that `build.props`
exist.