[Xamarin.ProjectTools] only save packages.config if it changes (#1901)
Context: https://github.com/xamarin/xamarin-android/pull/1878

The way `Xamarin.ProjectTools` saves `packages.config`, it would
accidentally trigger `_CleanIntermediateIfNuGetsChange`. It saves the
file *every* time, regardless if the contents changed or not. This
would play havoc with `_CleanIntermediateIfNuGetsChange` and cause it
to run when it shouldn't.

So I modeled what we are currently doing for conditionally saving the
project file. I keep the last `packages.config` contents around, and
make sure it only saves to disk if it changes. Using
`DateTimeOffSet.MinValue` as a `Timestamp` will prevent the file from
being written to disk.

I updated two tests to validate this change, which also adds some
additional validation to the `_CleanIntermediateIfNuGetsChange`
target.