[CodeBehind] Fix code generation for layouts with non-element root (#1889)
If the Android layout file being processed has an XML node other than
an element (e.g. a comment) directly following the XML declaration
node, the generator fails to properly locate the root element and
will miss any namespace declarations on it, as well as the
`/*/@xamarin:classes` attribute.

This results in failure to generate partial class code and will also
cause an error to parse any elements with the
`//*/@xamarin:managedType` attribute later in the layout file.

The fix is to skip over non-element nodes until we find the first
element, and then attempt to read namespaces and other attributes.